Tried rather unsuccessfully to get something going
with `specs2`... I was going to try to learn some basic testing and such
in Scala with a FizzBuzz project before diving into data structures but
after several attempts to use the Scala Build Tool `sbt` and just
vanilla `scala` \ `scalac` with `specs2` I find that I must throw in the
towel for the night. Gilli, will you show me how it's done? ;)

I had read a part of a book on Scala before but in a situation in which
I did not have the ability to actually try it out. This is my first
hands-on attempt with Scala.